python_engine/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pandas
openpyxl
supabase
requests
requests-cache
aiohttp
cachetools
beautifulsoup4
jinja2
python-multipart
playwright
playwright-stealth
urllib3
yfinance
numpy
alpaca-py>=0.17.0
black==24.1.0
pytest-asyncio
